pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["MaggeSN <122047522+MaggeSN@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"
pydantic = "^2.6.2"
numpy = "^1.26.4"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
flask = "^3.0.2"
python-keycloak = "^3.9.1"
python-decouple = "^3.8"
httpx = "^0.27.0"
pytest = "^8.1.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"





